operators

+++ src/Simplex/File.cpp
@@ -34,9 +34,9 @@ void Simplex::File::WriteOnFile(std::string path, std::string data)
     out_file.close();
 }
 
-std::vector<long> Simplex::File::GetIntsFromStringFile(std::string&& data_string)
+std::vector<long long> Simplex::File::GetIntsFromStringFile(std::string&& data_string)
 {
-    std::vector<long> data_int = {};
+    std::vector<long long> data_int = {};
 
     std::regex r("(\\+|-)?[[:digit:]]+");
     std::smatch m;
